Split discovery auto-config tests by scope
Add dedicated integration tests for blocking and reactive discovery auto-configuration to verify real CompositeDiscoveryClient wiring and service exposure with simple discovery instances. Refactor the existing tests to extend AutoConfigurationTest so they focus on conditional auto-configuration behavior (enabled/disabled properties and missing class scenarios) instead of full context integration.
